requirements.txt
beautifulsoup4==4.9.2
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
cycler==0.10.0
decorator==4.4.2
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
Pillow==7.2.0
pony==0.7.13
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==12.8
reportlab==3.5.52
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
urllib3==1.25.10
